Use smaller return type sizes for functions P_BoxOnLineSide, P_PointOnLineSide and P_PointOnDivlineSide

+++ p_local.h
@@ -142,10 +142,10 @@ extern intercept_t *intercept_p;
 typedef boolean (*traverser_t)(intercept_t *in);
 
 fixed_t P_AproxDistance(fixed_t dx, fixed_t dy);
-int P_PointOnLineSide(fixed_t x, fixed_t y, line_t *line);
-int P_PointOnDivlineSide(fixed_t x, fixed_t y, divline_t *line);
+byte P_PointOnLineSide(fixed_t x, fixed_t y, line_t *line);
+byte P_PointOnDivlineSide(fixed_t x, fixed_t y, divline_t *line);
 fixed_t P_InterceptVector(divline_t *v2, divline_t *v1);
-int P_BoxOnLineSide(fixed_t *tmbox, line_t *ld);
+byte P_BoxOnLineSide(fixed_t *tmbox, line_t *ld);
 
 extern fixed_t opentop;
 extern fixed_t openbottom;
